+++ e2e/CoValues/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @jazz-e2e/covalues
 
+## 0.0.85
+
+### Patch Changes
+
+-   Updated dependencies [c3f4e6b]
+-   Updated dependencies [d9152ed]
+    -   jazz-tools@0.8.5
+    -   cojson@0.8.5
+    -   jazz-react@0.8.5
+
 ## 0.0.84
 
 ### Patch Changes