+++ examples/pets/CHANGELOG.md
@@ -1,5 +1,15 @@
 # jazz-example-pets
 
+## 0.0.104
+
+### Patch Changes
+
+-   Updated dependencies [c3f4e6b]
+-   Updated dependencies [d9152ed]
+    -   jazz-tools@0.8.5
+    -   jazz-browser-media-images@0.8.5
+    -   jazz-react@0.8.5
+
 ## 0.0.103
 
 ### Patch Changes