+++ examples/todo/CHANGELOG.md
@@ -1,5 +1,14 @@
 # jazz-example-todo
 
+## 0.0.103
+
+### Patch Changes
+
+-   Updated dependencies [c3f4e6b]
+-   Updated dependencies [d9152ed]
+    -   jazz-tools@0.8.5
+    -   jazz-react@0.8.5
+
 ## 0.0.102
 
 ### Patch Changes