+++ packages/cojson-transport-ws/CHANGELOG.md
@@ -1,5 +1,13 @@
 # cojson-transport-nodejs-ws
 
+## 0.8.5
+
+### Patch Changes
+
+-   Updated dependencies [c3f4e6b]
+-   Updated dependencies [d9152ed]
+    -   cojson@0.8.5
+
 ## 0.8.3
 
 ### Patch Changes