+++ packages/jazz-react-auth-clerk/CHANGELOG.md
@@ -1,5 +1,16 @@
 # jazz-browser-media-images
 
+## 0.8.5
+
+### Patch Changes
+
+-   Updated dependencies [c3f4e6b]
+-   Updated dependencies [d9152ed]
+    -   jazz-tools@0.8.5
+    -   cojson@0.8.5
+    -   jazz-browser-auth-clerk@0.8.5
+    -   jazz-react@0.8.5
+
 ## 0.8.3
 
 ### Patch Changes